+++ frontend/components/Domain/Recipe/RecipeContextMenu.vue
@@ -2,8 +2,6 @@
   <div class="text-center">
     <!-- Recipe Share Dialog -->
     <RecipeDialogShare v-model="shareDialog" :recipe-id="recipeId" :name="name" />
-    <!-- I Made This Dialog -->
-    <RecipeDialogMadeThis v-model="madeThisDialog" :recipe-slug="slug" />
     <BaseDialog
       v-model="recipeDeleteDialog"
       :title="$t('recipe.delete-recipe')"
@@ -101,7 +99,6 @@
 <script lang="ts">
 import { defineComponent, reactive, toRefs, useContext, useRouter, ref } from "@nuxtjs/composition-api";
 import RecipeDialogShare from "./RecipeDialogShare.vue";
-import RecipeDialogMadeThis from "./RecipeDialogMadeThis.vue";
 import { useUserApi } from "~/composables/api";
 import { alert } from "~/composables/use-toast";
 import { planTypeOptions } from "~/composables/use-group-mealplan";
@@ -113,7 +110,6 @@ import { useCopy } from "~/composables/use-copy";
 export interface ContextMenuIncludes {
   delete: boolean;
   edit: boolean;
-  madeThis: boolean;
   download: boolean;
   mealplanner: boolean;
   shoppingList: boolean;
@@ -132,7 +128,6 @@ export interface ContextMenuItem {
 export default defineComponent({
   components: {
     RecipeDialogShare,
-    RecipeDialogMadeThis,
   },
   props: {
     useItems: {
@@ -140,7 +135,6 @@ export default defineComponent({
       default: () => ({
         delete: true,
         edit: true,
-        madeThis: true,
         download: true,
         mealplanner: true,
         shoppingList: true,
@@ -202,7 +196,6 @@ export default defineComponent({
 
     const state = reactive({
       shareDialog: false,
-      madeThisDialog: false,
       recipeDeleteDialog: false,
       mealplannerDialog: false,
       shoppingListDialog: false,
@@ -230,12 +223,6 @@ export default defineComponent({
         icon: $globals.icons.delete,
         color: "error",
         event: "delete",
-      },
-      madeThis: {
-        title: "I Made This",
-        icon: $globals.icons.chefHat,
-        color: undefined,
-        event: "madeThis",
       },
       download: {
         title: i18n.tc("general.download"),
@@ -351,9 +338,6 @@ export default defineComponent({
         state.recipeDeleteDialog = true;
       },
       edit: () => router.push(`/recipe/${props.slug}` + "?edit=true"),
-      madeThis: () => {
-        state.madeThisDialog = true;
-      },
       download: handleDownloadEvent,
       mealplanner: () => {
         state.mealplannerDialog = true;
